Firestoreに関する型や関数を削除
