+++ app/utils/firebase.ts
@@ -1,7 +1,6 @@
 import { initializeApp } from 'firebase/app'
 import { connectAuthEmulator, getAuth } from 'firebase/auth'
 import { connectFirestoreEmulator, getFirestore } from 'firebase/firestore'
-import { getAnalytics } from 'firebase/analytics'
 
 const app = initializeApp({
   apiKey: 'AIzaSyAl93UHoFUQb1wfPhifDNPX1UDZpheWfxI',
