+++ server/utils/database.ts
@@ -35,3 +35,15 @@ export function parseChallengeResult<T extends HasJudgeStatus>(
     status: challengeResult.status as JudgeStatus,
   }
 }
+
+/**
+ * 20桁の英数字のID（Firestoreの自動生成IDと同じ形式）を作成する
+ */
+export function generateAutoId(): string {
+  const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
+  let autoId = ''
+  for (let i = 0; i < 20; i++) {
+    autoId += chars.charAt(Math.floor(Math.random() * chars.length))
+  }
+  return autoId
+}
